requirements.txt
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
fonttools==4.49.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.5.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
Telethon==1.34.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
